refactor: change post to post.view
