+++ resources/views/livewire/post/view.blade.php
